refactor: 날짜 검사 함수 분리
